+++ perl/FSM/Support/ISFPublicInterfaceContract.pm
@@ -435,6 +435,7 @@ sub build_isf_public_interface_contract {
             't/1266-isf-enum-member-drive-values.t',
             't/1267-isf-enum-member-drive-call-values.t',
             't/1268-isf-enum-member-drive-call-expression-values.t',
+            't/1269-isf-enum-member-actor-params.t',
         ],
         guidance => [
             'Treat this as the first bounded public ISF downstream-consumer contract, advertised through embedding.isf_public_interface.',